refactor junit runner and write as binary observation, need loading
