+++ src/main/java/fr/inria/diversify/compare/ObjectLog.java
@@ -1,5 +1,7 @@
 package fr.inria.diversify.compare;
 
+import java.io.FileOutputStream;
+import java.io.ObjectOutputStream;
 import java.lang.reflect.*;
 import java.util.HashMap;
 import java.util.Map;
@@ -12,22 +14,22 @@ import fr.inria.diversify.dspot.TypeUtils;
  * Time: 14:31
  */
 public class ObjectLog {
-    private static ObjectLog singleton ;
-    protected Map<String, Observation> observations;
-    protected MethodsHandler methodsHandler;
-    protected Invocator invocator;
-    protected int maxDeep = 4;
-    protected Map<String, Object> objects;
+    private static ObjectLog singleton;
+    private Map<String, Observation> observations;
+    private MethodsHandler methodsHandler;
+    private Invocator invocator;
+    private int maxDeep = 4;
+    private Map<String, Object> objects;
 
     private ObjectLog() {
-        this.observations = new HashMap<String, Observation>();
-        this.objects = new HashMap<String, Object>();
+        this.observations = new HashMap<>();
+        this.objects = new HashMap<>();
         this.methodsHandler = new MethodsHandler(true, true);
         this.invocator = new Invocator(1);
     }
 
     protected static ObjectLog getSingleton() {
-        if(singleton == null) {
+        if (singleton == null) {
             singleton = new ObjectLog();
         }
         return singleton;
@@ -43,20 +45,20 @@ public class ObjectLog {
     }
 
     public static void logObject(Object object, String stringObject, String positionId) {
-        if(object != null) {
+        if (object != null) {
             getSingleton().objects.put(stringObject, object);
         }
     }
 
     public void pLog(Object object, String stringObject, String positionId, int deep) {
-        if(deep < maxDeep) {
+        if (deep < maxDeep) {
             if (object == null) {
                 addObservation(positionId, stringObject, null);
             } else if (TypeUtils.isPrimitive(object)) {
                 addObservation(positionId, stringObject, object);
             } else if (TypeUtils.isPrimitiveArray(object)) {
                 addObservation(positionId, stringObject, object);
-            } else if(TypeUtils.isPrimitiveCollectionOrMap(object)) {
+            } else if (TypeUtils.isPrimitiveCollectionOrMap(object)) {
 //                typeOfIterable(object);
                 addObservation(positionId, stringObject, object);
             } else {
@@ -67,7 +69,7 @@ public class ObjectLog {
     }
 
     protected void addObservation(String positionId, String stringObject, Object value) {
-        if(!observations.containsKey(positionId)) {
+        if (!observations.containsKey(positionId)) {
             observations.put(positionId, new Observation());
         }
         observations.get(positionId).add(stringObject, value);
@@ -82,11 +84,12 @@ public class ObjectLog {
                     }
                 }
             }
-        } catch (Exception e) {}
+        } catch (Exception e) {
+        }
     }
 
-    protected void observeNotNullObject(Object o,  String stringObject, String positionId, int deep) {
-        if(deep < maxDeep) {
+    protected void observeNotNullObject(Object o, String stringObject, String positionId, int deep) {
+        if (deep < maxDeep) {
             for (Method method : methodsHandler.getAllMethods(o)) {
                 Invocation invocation = new Invocation(o, method);
                 invocator.invoke(invocation);
@@ -100,7 +103,24 @@ public class ObjectLog {
         }
     }
 
+    public static void writeObservationToFile() {
+        try {
+            FileOutputStream fos = new FileOutputStream("observations.ser");
+            try (ObjectOutputStream oos = new ObjectOutputStream(fos)) {
+                oos.writeObject(singleton.observations);
+            } catch (Exception e) {
+                throw new RuntimeException(e);
+            }
+        } catch (Exception e) {
+            throw new RuntimeException(e);
+        }
+    }
+
     public static Map<String, Observation> getObservations() {
+        return singleton.observations;
+    }
+
+    public static Map<String, Observation> buildObservationFromFile() {
         return singleton.observations;
     }
 }
